Initial commit for wireless message passing via socket programming. Tested with Python 3.6.8. No additional modules required.
-'server.py' should be running on the ATM (server). Ensure this is running before making client calls. Will display a custom pop-up error message
-'client.py' should be imported into another script to signal a pop-up error message on the ATM (server).
-'test.py' is an example test script on how to use client.py.

NOTE: The IP address for server.py and client.py should be the same, e.g. match the IP address of the ATM (server.py should print out the IP address on start-up as an aid).

Set the USE_LOCAL_HOST flag to True in server.py and client.py to test on the same machine.